+++ src/format/html/elements/division.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace VaTools\Format\Html\Elements;
+
+use \VaTools\Format\Html\HtmlElement;
+
+class HtmlDivision extends HtmlElement {
+
+    /**
+     * @param string $text Inner text
+     */
+    public function __construct(string $text = '') {
+
+        parent::__construct('div', $text);
+    }
+}