+++ src/format/html/inc.php
@@ -1,0 +1,9 @@
+<?php
+
+use \Vosiz\Utils\Io;
+
+require_once(__DIR__.'/exc.php');
+require_once(__DIR__.'/core/element.php');
+
+Io\Inc::Dir(__DIR__.'/base');
+Io\Inc::Dir(__DIR__.'/elements');